Separate tty logic from vbe.cpp
