+++ arch/x86_64/kernel/kernel.cpp
@@ -33,12 +33,8 @@ void write_ff_info() {
 
 
 void kernel_main() {
-    write_ff_info();
-    for (int i = 0; i < 47; i++)
-        printf("test %d\n", i);
-
-
-    printf("Scrolling demo complete\n");
-    trace::panic("Test");
+    for (int i = 0; i < 58; i++)
+        printf("I = %d\n", i);
+    trace::panic("Reached the end of kernel");
 }
 }  // namespace firefly::kernel::main
