+++ include/x86_64/kernel.hpp
@@ -2,4 +2,5 @@
 
 namespace firefly::kernel::main {
 void kernel_main();
+void write_ff_info();
 }
