+++ include/x86_64/libk++/vector_math.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <stdint.h>
+
+namespace firefly::libkern {
+typedef struct
+{
+    uint64_t x;
+    uint64_t y;
+} Vec2;
+}  // namespace firefly::libkern
