--- include/x86_64/memory-manager/bootstrap-buddy.hpp
@@ -1,8 +1,0 @@
-// #pragma once
-// #include <stl/cstdlib/cstdint.h>
-
-// #include <x86_64/multiboot2.hpp>
-
-// namespace firefly::mm::buddy {
-// void bootstrap_buddy(multiboot_uint64_t &mmap_base_entry, multiboot_uint64_t &length);
-// }  // namespace firefly::mm::buddy
